# docs/upgrade/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReVector Core Engine",
    "description": "True editable geometry. No raster wrappers. Deploy behind JerseyOS authentication.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/revector/correct-geometry": {
      "post": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Geometry",
        "operationId": "geometry_api_revector_correct_geometry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeometryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/segment": {
      "post": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Segment",
        "operationId": "segment_api_revector_segment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/export": {
      "post": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Export",
        "operationId": "export_api_revector_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/revector/jobs/{job_id}": {
      "get": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Job",
        "operationId": "job_api_revector_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "processing jobs"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_revector_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEvent": {
        "properties": {
          "error_id": {
            "type": "string",
            "title": "Error Id"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "part_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "technical_summary": {
            "type": "string",
            "title": "Technical Summary"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "recoverable": {
            "type": "boolean",
            "title": "Recoverable"
          },
          "suggested_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Actions"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "error_id",
          "error_code",
          "code",
          "category",
          "phase",
          "message",
          "technical_summary",
          "retryable",
          "recoverable",
          "timestamp"
        ],
        "title": "ErrorEvent"
      },
      "ExportRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "formats": {
            "items": {
              "type": "string",
              "enum": [
                "svg",
                "pdf",
                "eps",
                "png",
                "zip"
              ]
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Formats"
          },
          "part_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 100,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "ExportRequest"
      },
...
      "GeometryRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "corners": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array",
                "maxItems": 4,
                "minItems": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Corners"
          },
          "auto": {
            "type": "boolean",
            "title": "Auto",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "GeometryRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "job_state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Job State"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEvent"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "process_event": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process Event"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "project_id",
          "stage",
          "status",
          "job_state",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse"
      },
      "ManualPart": {
        "properties": {
          "polygon": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "maxItems": 10000,
            "minItems": 3,
            "title": "Polygon"
          },
          "type": {
            "type": "string",
            "enum": [
              "front_body",
              "back_body",
              "left_sleeve",
              "right_sleeve",
              "left_shoulder",
              "right_shoulder",
              "front_collar",
              "back_collar",
              "left_cuff",
              "right_cuff",
              "trim",
              "top_trim",
              "bottom_trim",
              "other_part",
              "unknown"
            ],
            "title": "Type",
            "default": "unknown"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name",
            "default": "Manual part"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "polygon"
        ],
        "title": "ManualPart"
      },
...
      "SegmentRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "parts": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ManualPart"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Parts"
          },
          "points": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array",
                "maxItems": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "SegmentRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}